+++ src/mod/Gloabl.h
@@ -1,12 +1,4 @@
 #pragma once
-#include "Config/Config.h"
-#include "I18n/I18n.h"
-#include "ll/api/io/Logger.h"
 
-
-
-namespace my_mod {
-extern ll::io::Logger& logger;
-extern Config&         config;
-
-}
+// Backward-compatible include for old misspelled path.
+#include "mod/Global.h"