+++ src/mod/Global.h
@@ -1,0 +1,12 @@
+#pragma once
+
+#include "Config/Config.h"
+#include "I18n/I18n.h"
+#include "ll/api/io/Logger.h"
+
+namespace my_mod {
+
+extern ll::io::Logger& logger;
+extern Config&         config;
+
+} // namespace my_mod